chore(deps): upgrade angular in examples/carbon accounting
Upgraded both the supply chain app and the carbon accounting app GUIs
to Angular v13 so that the vulnerability in question is resolved through
the transient dependencies of Angular development tooling related
packages.

Also upgraded the dependency @angular-builders/custom-webpack which was
responsible for the transient dependency with the vulnerability.

Fixes #1651

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>