Fabric Test networks and simple state chaincode for Weaver Data Sharing example
Imported Fabric testnets and 'simplestate' chaincode (which simply manages key-value pairs on ledger) to the examples folder.
Added placeholder instructions in a package.json to start the testnets and deploy the chaincode.
Need to integrate it with other Weaver dependencies and run an end-to-end example.
Need to then refactor the pipeline for integration with Cactus.

Signed-off-by: VRamakrishna <vramakr2@in.ibm.com>