fix(connector-fabric): runTransactionV1 Open API validation crash
The "express-openapi-validator" package was crashing while trying to
parse our Fabric connector's Open API spec because of the typo that
in the specs the "description" field for some schema elements were named
"properties" which is a reserved field in the schema of the specs.

Fixing the typo in the spec document made the problem go away.