refactor(connector-ethereum): make web3 race condition fix less stateful
Does the same thing as the fix provided before, but with less management
of state because it attaches boolean variable to the socket objects of
Web3.

Riskier in terms of settting variables on objects of third party libraries
that we don't control, but at the same time less risky of having additional
bugs in it because it's much less logic.

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>