feat(plugin-keychain-memory): add REST API endpoint implementations
BREAKING CHANGE: Some of the schema elements that were redundant
got removed from the openapi.json spec of the memory plugin.
The same schemas are now being pulled from the core-api package.
In this sense there is no breaking change because these two
schemas were already equivalent anyway, but the fact is
the openapi.json had schema elements removed from it anyway
so this is something that people need to be aware of as they
upgrade to v2.0.0 in the near future.

1. Added the REST API endpoints to the plugin so that it is accessible over
HTTP for testing/demo purposes as necessary.
2. Also refactored the openapi.json spec file so that it pulls the schema
definitions from the core-api package the same way as the
keychain-memory-wasm plugin is doing already.

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>
(cherry picked from commit 8d0b9ac5d7fa282b3de7b9db4caab7710b879747)