refactor(connector-besu): upgrade to Web3JS, use bigints as strings
WORK IN PROGRESS - BUILD IS CURRENTLY BROKEN

**BREAKING CHANGE**

1. On the surface it's just a simple dependency upgrade, but unfortunately
we've made the mistake earlier to use Web3's own types as the return types
for some of the public facing methods and because of this there is breaking
changes to be made.
2. On the OpenAPI specification level it is also necessary to make breaking
changes because we have to deal with int64s (BigInt type in Javascript)
which cannot be represented in JSON and therefore we must juggle the transport
of these values as strings that get converted back and forth between BigInt
and string.

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>