docs(examples/supply-chain-app): enable local source builds in container
WORK IN PROGRESS

The problem is that the image is 11 GB large.
We need to write a script that cleans out most of the packages that
aren't being used by the supply chain example.

If ESBuild or Vite could bundle it the way Webpack is supposed to that
would be even better but this is 99.99999% not going to work out of the box
and is most likely a very deep rabbithole.

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>