feat(core-api): add p2p messaging inbox, outbox for plugin interface
Huge diff, apologies. The change itself is very simple, it adds a couple
methods to the plugin interface definition so that plugins can send and
receive messages to and from other plugins via the API servers peer to
peer messaging.

1. The inbox is an RxJS subjct where the API server can place messages
delivered from other plugins and
2. The outbox is an RxJS observable that the API server can use to
subscribe to new outgoing messages being sent by the plugin.
3. Currently there is no implementation behind this at all, but the skeleton
plumbing is being put in place so that a later change can introduce it.

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>