fix(connector-ethereum): validation of Open API HTTP request types
The tests are passing because they aren't covering the use-case of
the plugin being loaded into the API server which automatically wires
up an OpenAPI HTTP request validation middleware (which the tests don't do)
and thus the validation is not being exercised.

For example the test case at:
connector-ethereum/src/test/typescript/integration/geth-transact-and-gas-fees.test.ts
is not exercising the validation middleware because it is not being loaded into an API server.
We can simulate the API server scenario by leveraging the re-usable function
of the core package that exports the logic to wire up the validation middleware.
This will allow us to exercise the validation logic in a more realistic scenario.

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>